boot-selector/requirements.txt
aiomqtt>=2.0.0
python-dotenv>=1.0.0
tftpy>=0.8.2
